Custom CatalogFileIndex (#364)
