+++ sql/catalyst/src/main/scala/org/apache/spark/sql/internal/StaticSQLConf.scala
@@ -36,6 +36,12 @@ object StaticSQLConf {
     .stringConf
     .createWithDefault(Utils.resolveURI("spark-warehouse").toString)
 
+  val CATALOG_FILE_INDEX_IMPLEMENTATION =
+    buildStaticConf("spark.sql.catalogFileIndexImplementation")
+      .internal()
+      .stringConf
+      .createWithDefault("hive")
+
   val CATALOG_IMPLEMENTATION = buildStaticConf("spark.sql.catalogImplementation")
     .internal()
     .version("2.0.0")
