+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/CatalogFileIndex.scala
@@ -17,92 +17,60 @@
 
 package org.apache.spark.sql.execution.datasources
 
-import java.net.URI
+import scala.util.control.NonFatal
 
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.Path
-
+import org.apache.spark.SparkConf
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.catalog.CatalogTable
 import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.types.StructType
-
+import org.apache.spark.sql.internal.StaticSQLConf.CATALOG_FILE_INDEX_IMPLEMENTATION
+import org.apache.spark.util.Utils
 
 /**
  * A [[FileIndex]] for a metastore catalog table.
- *
- * @param sparkSession a [[SparkSession]]
- * @param table the metadata of the table
- * @param sizeInBytes the table's data size in bytes
  */
-class CatalogFileIndex(
-    sparkSession: SparkSession,
-    val table: CatalogTable,
-    override val sizeInBytes: Long) extends FileIndex {
-
-  protected val hadoopConf: Configuration = sparkSession.sessionState.newHadoopConf()
-
-  /** Globally shared (not exclusive to this table) cache for file statuses to speed up listing. */
-  private val fileStatusCache = FileStatusCache.getOrCreate(sparkSession)
-
-  assert(table.identifier.database.isDefined,
-    "The table identifier must be qualified in CatalogFileIndex")
-
-  private val baseLocation: Option[URI] = table.storage.locationUri
-
-  override def partitionSchema: StructType = table.partitionSchema
-
-  override def rootPaths: Seq[Path] = baseLocation.map(new Path(_)).toSeq
-
-  override def listFiles(
-      partitionFilters: Seq[Expression], dataFilters: Seq[Expression]): Seq[PartitionDirectory] = {
-    filterPartitions(partitionFilters).listFiles(Nil, dataFilters)
-  }
-
-  override def refresh(): Unit = fileStatusCache.invalidateAll()
+trait CatalogFileIndex extends FileIndex {
 
   /**
-   * Returns a [[InMemoryFileIndex]] for this table restricted to the subset of partitions
+   * Returns a [[FileIndex]] for this table restricted to the subset of partitions
    * specified by the given partition-pruning filters.
    *
    * @param filters partition-pruning filters
    */
-  def filterPartitions(filters: Seq[Expression]): InMemoryFileIndex = {
-    if (table.partitionColumnNames.nonEmpty) {
-      val startTime = System.nanoTime()
-      val selectedPartitions = sparkSession.sessionState.catalog.listPartitionsByFilter(
-        table.identifier, filters)
-      val partitions = selectedPartitions.map { p =>
-        val path = new Path(p.location)
-        val fs = path.getFileSystem(hadoopConf)
-        PartitionPath(
-          p.toRow(partitionSchema, sparkSession.sessionState.conf.sessionLocalTimeZone),
-          path.makeQualified(fs.getUri, fs.getWorkingDirectory))
-      }
-      val partitionSpec = PartitionSpec(partitionSchema, partitions)
-      val timeNs = System.nanoTime() - startTime
-      new InMemoryFileIndex(sparkSession,
-        rootPathsSpecified = partitionSpec.partitions.map(_.path),
-        parameters = Map.empty,
-        userSpecifiedSchema = Some(partitionSpec.partitionColumns),
-        fileStatusCache = fileStatusCache,
-        userSpecifiedPartitionSpec = Some(partitionSpec),
-        metadataOpsTimeNs = Some(timeNs))
-    } else {
-      new InMemoryFileIndex(sparkSession, rootPaths, parameters = table.storage.properties,
-        userSpecifiedSchema = None, fileStatusCache = fileStatusCache)
+  def filterPartitions(filters: Seq[Expression]): FileIndex
+
+}
+
+trait CatalogFileIndexFactory {
+
+  /**
+   * Creates [[CatalogFileIndex]] for given table
+   */
+  def create(
+    spark: SparkSession,
+    catalogTable: CatalogTable,
+    tableSize: Long): CatalogFileIndex
+
+}
+
+object CatalogFileIndexFactory {
+
+  def reflect[T <: CatalogFileIndexFactory](conf: SparkConf): T = {
+    val className = fileIndexClassName(conf)
+    try {
+      val ctor = Utils.classForName(className).getDeclaredConstructor()
+      ctor.newInstance().asInstanceOf[T]
+    } catch {
+      case NonFatal(e) =>
+        throw new IllegalArgumentException(s"Error while instantiating '$className':", e)
     }
   }
 
-  override def inputFiles: Array[String] = filterPartitions(Nil).inputFiles
-
-  // `CatalogFileIndex` may be a member of `HadoopFsRelation`, `HadoopFsRelation` may be a member
-  // of `LogicalRelation`, and `LogicalRelation` may be used as the cache key. So we need to
-  // implement `equals` and `hashCode` here, to make it work with cache lookup.
-  override def equals(o: Any): Boolean = o match {
-    case other: CatalogFileIndex => this.table.identifier == other.table.identifier
-    case _ => false
+  private def fileIndexClassName(conf: SparkConf): String = {
+    conf.get(CATALOG_FILE_INDEX_IMPLEMENTATION) match {
+      case "hive" => "org.apache.spark.sql.execution.datasources.HiveCatalogFileIndexFactory"
+      case name => name
+    }
   }
 
-  override def hashCode(): Int = table.identifier.hashCode()
 }
